README.md
A project where i design and build my own simple smartwatch. It not intended to be a super complex product, offering little more functionality than a standard analog watch. In this project I design the UI in Figma, aquire the necessary hardware, code it in the Arduino IDE with assistance from Squareline Studios, and design the casing for the watch itself.